fix bug

+++ setup.sh
@@ -1,17 +1,27 @@
-echo "Dinh Truong xin chao anh em!"
-echo "Setting up the environment!"
+
+echo "[+] Setup environment"
 echo "=========================="
+start_time = $(date +%s)
+
 pip install -q -U datasets
-pip install -q -r requirements.txt
-pip install -q -r /content/code-generation-codet5p/requirements.txt
-pip install -q -r /kaggle/working/code-generation-codet5p/requirements.txt
-echo "Installed packages:"
-echo "- datasets"
-echo "- wandb"
-echo "- huggingface_hub"
-echo "- accelerate"
-echo "- peft"
-echo "- bitsandbytes"
-echo "- human-eval"
+if [[ -f "requirements.txt" ]]; then
+    pip install -q -r requirements.txt
+fi
+
+if [[ -f "/content/code-generation-codet5p/requirements.txt"]]; then
+    pip install -q -r /content/code-generation-codet5p/requirements.txt
+fi
+
+if [[ -f "/kaggle/working/code-generation-codet5p/requirements.txt"]]; then
+    pip install -q -r /kaggle/working/code-generation-codet5p/requirements.txt
+
 echo "=========================="
-echo "Set up complete!"
+echo "[+] Finish setup environment"
+
+end_time=$(date +%s)
+setup_time=$((end_time - start_time))
+
+mins = $((setup_time / 60))
+secs = $((setup_time % 60))
+
+echo "Time taken for setup: $mins mins $secs secs."
